dom/tests/mochitest/webcomponents/test_shadowroot_inert_element.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=806506
-->
<head>
  <title>Test for inert elements in ShadowRoot</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="runChecks();">
<div id="grabme"></div>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=806506">Bug 806506</a>
<script>

var element = document.getElementById("grabme");
var shadow = element.createShadowRoot();

// Check that <base> is inert.
shadow.innerHTML = '<base href="http://www.example.org/" />';
isnot(document.baseURI, "http://www.example.org/", "Base element should be inert in ShadowRoot.");

SimpleTest.waitForExplicitFinish();

SimpleTest.finish();

</script>
</body>
</html>
